+++ app/lib/ui/components/control_item/control_item_enum.dart
@@ -12,6 +12,6 @@
 
 enum ControlItemType {
   checkbox,
-  toggleSwitch,
+  switchButton,
   radioButton,
 }